iot/p9-iot-db/esp32/requirements.txt
flask==3.1.2
flask-mysql-connector==1.1.0
python-dotenv==1.0.1
flask-cors==4.0.1
requests==2.32.3